Implement GPU data reordering and update test suite to Float64
- Implemented physical data reordering and reverse force reordering in MollyCUDAExt.jl to improve cache locality.
- Enabled tile classification (CLEAN vs EXCLUDED) in the interacting block kernel.
- Added fast bitwise implementations for Morton code generation in 2D and 3D.
- Updated GPU test suites to use Float64 and tightened CPU-GPU comparison tolerances to 1e-10.
- Fixed NoNeighborFinder usage and interaction setup in test/gpu_tile_lists.jl.

+++ test/constraints.jl
@@ -306,7 +306,7 @@ end
 
 @testset "Constraints protein CPU/GPU" begin
     pdb_fp = joinpath(data_dir, "1ubq.pdb") # No solvent
-    T = Float32
+    T = Float64
     ff = MolecularForceField(
         T,
         joinpath(ff_dir, "ff99SBildn.xml"),
